add path check to prevent deleting everything

+++ App/metadata_handler.cpp
@@ -185,9 +185,13 @@ void MetadataHandler::FindBinariesInFolder(){
 
 void MetadataHandler::DeleteExtractedFolder(){
     //We are now done with the extracted directory that we made. We should delete it to avoid any issues
-    QDir dir(extract_path_);
-    dir.removeRecursively();
-    extract_path_ = "";
+    //If extract path is "", it can delete a lot of files unexpectedly! It should be impossible to wind up
+    //in that case, but checking to be extra careful
+    if(extract_path_ != ""){
+        QDir dir(extract_path_);
+        dir.removeRecursively();
+        extract_path_ = "";
+    }
 }
 
 bool MetadataHandler::CheckHardwareAndElectronics(){
